Switch storage from SQL Server to PostgresSQL for Hangfire
The commit switches the Hangfire storage from SQL Server to PostgreSQL by updating the Hangfire dependencies in Project.cs and replacing the existing UseSqlServerStorage method with the UsePostgreSqlStorage. Updated configurations accordingly in appsettings.json and made HangfireDBConnection use the new PostgreSQL. Also updated the launchSettings to disable default browser launch for the project.