Add user verification via generated code
A new code generator class is added, which generates random six-digit codes. These codes are used for user verification in TelegramBot. Relevant changes are also made in "HandleUpdateAsync" method where the code checks if the user exists and if not, returns a message. Otherwise, it checks for an existing verification code, if there isn't any, it generates a new one for the user, saves it in db, and sends it to the user. If there is an existing unexpired code, it simply sends the existing code to the user.