Remove TgBotDbContext and add UserChatPayment model
The TgBotDbContext was removed as it is no longer necessary. A new model called UserChatPayment was added to ApplicationIdentityDbContext. In the TelegramBot class, the functionality to check for an existing user based on Telegram ID was added before starting the specific command. Several other code adjustments were done to provide these changes.