Inject NotificationService into SubscriptionCard.razor
The `NotificationService` has been injected into the `SubscriptionCard.razor` file to handle user notifications. Notifications have been added in the `if` and `else` blocks of the `currentPlan` check to inform the user about successful subscription to the free plan. The notifications are displayed based on whether the current plan is null or not, after the `HttpClient` posts the new user subscription to the specified URL.